L-8-6 Edit + Delete

+++ index.view.php
@@ -39,7 +39,7 @@
       <td><a href="/show.php?id=<?=$post['id'];?>"><?=$post['title'];?></a></td>
       <td>
       	<a href="/edit.php?id=<?=$post['id'];?>" class="btn btn-warning">Edit</a>
-      	<a href="/delete.php?id=<?=$post['id'];?>" class="btn btn-danger">Delete</a>
+      	<a href="/delete.php?id=<?=$post['id'];?>" class="btn btn-danger" onclick="return confirm('are you sure?');"  >Delete</a>
       </td>
     </tr>
   	<?php endforeach; ?>
